Split factory helpers
