+++ packages/factory/src/helpers/logger.ts
@@ -2,22 +2,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-import { ALL_ENVS } from '@dd/core/constants';
 import { formatDuration } from '@dd/core/helpers/strings';
-import type {
-    BuildReport,
-    BundlerFullName,
-    BundlerName,
-    Env,
-    FactoryMeta,
-    GetLogger,
-    GlobalContext,
-    LogLevel,
-    Options,
-    OptionsWithDefaults,
-    TimeLog,
-    Timer,
-} from '@dd/core/types';
+import type { BuildReport, GetLogger, LogLevel, TimeLog, Timer } from '@dd/core/types';
 import c from 'chalk';
 
 const logPriority: Record<LogLevel, number> = {
@@ -161,77 +147,3 @@ export const getLoggerFactory =
             debug: (text: any) => log(text, 'debug'),
         };
     };
-
-export const getContext = ({
-    options,
-    bundlerName,
-    bundlerVersion,
-    version,
-}: {
-    options: OptionsWithDefaults;
-    bundlerName: BundlerName;
-    bundlerVersion: string;
-    version: FactoryMeta['version'];
-}): GlobalContext => {
-    const cwd = process.cwd();
-    const variant = bundlerName === 'webpack' ? bundlerVersion.split('.')[0] : '';
-    const build: BuildReport = {
-        errors: [],
-        warnings: [],
-        logs: [],
-        timings: [],
-        bundler: {
-            name: bundlerName,
-            fullName: `${bundlerName}${variant}` as BundlerFullName,
-            variant,
-            version: bundlerVersion,
-        },
-    };
-
-    // Use "production" if there is no env passed.
-    const passedEnv: Env = (process.env.BUILD_PLUGINS_ENV as Env) || 'production';
-    // Fallback to "development" if the passed env is wrong.
-    const env: Env = ALL_ENVS.includes(passedEnv) ? passedEnv : 'development';
-    const context: GlobalContext = {
-        auth: options.auth,
-        pluginNames: [],
-        bundler: {
-            ...build.bundler,
-            // This will be updated in the bundler-report plugin once we have the configuration.
-            outDir: cwd,
-        },
-        build,
-        // This will be updated in the bundler-report plugin once we have the configuration.
-        cwd,
-        env,
-        getLogger: getLoggerFactory(build, options.logLevel),
-        // This will be updated in the injection plugin on initialization.
-        asyncHook: () => {
-            throw new Error('AsyncHook function called before it was initialized.');
-        },
-        hook: () => {
-            throw new Error('Hook function called before it was initialized.');
-        },
-        // This will be updated in the injection plugin on initialization.
-        inject: () => {
-            throw new Error('Inject function called before it was initialized.');
-        },
-        sendLog: () => {
-            throw new Error('SendLog function called before it was initialized.');
-        },
-        plugins: [],
-        start: Date.now(),
-        version,
-    };
-
-    return context;
-};
-
-export const validateOptions = (options: Options = {}): OptionsWithDefaults => {
-    return {
-        auth: {},
-        disableGit: false,
-        logLevel: 'warn',
-        ...options,
-    };
-};
